Initial release
Wheel Of Fortune Sound Player for iOS

This is a small, standalone iOS app. It plays some common Wheel of
Fortune sound effects for use with a Wheel of Fortune game.

Note: This was created for use at our local high school for the senior
party Wheel of Fortune room. We have a large wheel, and the kids come
in and play to win prizes. The sound effects add to the ambience of the
room.

This small app was built in Xcode, and compiles and runs on iOS 8 or
newer.

Enjoy!

Fork and improve, if you are so inclined.

TPM 2/20/15